Predictive distribution added
Made a separate file and put in all the helper math functions like
transpose, log beta, etc.
Added the predictive probability computation for new data.
Fixed validity check for the a priori matrix in the MAP function.
